+++ src/containers/HeaderContainer/HeaderContainer.vue
@@ -27,12 +27,7 @@
         :address="address"
       />
       <logout-modal ref="logout" />
-      <issue-log-modal
-        v-if="Object.keys.length > 0"
-        ref="issuelog"
-        :error="error"
-        :resolver="resolver"
-      />
+      <issue-log-modal ref="issuelog" :error="error" :resolver="resolver" />
 
       <!-- Desktop menu *********************************** -->
       <div class="fixed-header-wrap">
@@ -228,7 +223,7 @@
 </template>
 
 <script>
-import { mapState } from 'vuex';
+import { mapState, mapActions } from 'vuex';
 import store from 'store';
 import { Misc, Toast } from '@/helpers';
 import Blockie from '@/components/Blockie';
@@ -285,7 +280,8 @@ export default {
       'web3',
       'account',
       'gettingStartedDone',
-      'locale'
+      'locale',
+      'tempHide'
     ]),
     showButtons() {
       if (
@@ -346,11 +342,10 @@ export default {
     };
 
     this.$eventHub.$on('issueModal', (error, resolve) => {
-      // eslint-disable-next-line
       let errorPop = store.get('errorPop') || 0;
       errorPop += 1;
       store.set('errorPop', errorPop);
-      if (store.get('neverReport')) {
+      if (this.tempHide) {
         resolve(false);
       } else {
         this.$refs.issuelog.$refs.issuelog.show();
@@ -374,6 +369,7 @@ export default {
     this.$eventHub.$off('open-settings');
   },
   methods: {
+    ...mapActions('main', ['setLocale']),
     getCurrentLang() {
       const storedLocale = this.supportedLanguages.find(item => {
         return item.langCode === this.locale;
@@ -403,7 +399,7 @@ export default {
       this.$i18n.locale = obj.langCode;
       this.currentName = obj.name;
       this.currentFlag = obj.flag;
-      this.$store.dispatch('setLocale', obj.langCode);
+      this.setLocale(obj.langCode);
     },
     logout() {
       this.$refs.logout.$refs.logout.show();