+++ src/layouts/AccessWalletLayout/components/PrivateKeyModal/PrivateKeyModal.vue
@@ -47,7 +47,7 @@
 <script>
 import { WalletInterface } from '@/wallets';
 import { PRIV_KEY as privKeyType } from '@/wallets/bip44/walletTypes';
-import { mapState } from 'vuex';
+import { mapState, mapActions } from 'vuex';
 import { isHexString } from 'ethereumjs-util';
 import WarningMessage from '@/components/WarningMessage';
 import CustomerSupport from '@/components/CustomerSupport';
@@ -76,21 +76,19 @@ export default {
         : this.privateKey;
     }
   },
-  mounted() {},
   methods: {
+    ...mapActions('main', ['decryptWallet']),
     unlockWallet() {
       this.spinner = true;
-      this.$store
-        .dispatch('main/decryptWallet', [
-          new WalletInterface(this.actualPrivKey, false, privKeyType)
-        ])
-        .then(() => {
-          this.actualPrivKey = '';
-          this.spinner = false;
-          this.$router.push({
-            path: 'interface'
-          });
+      this.decryptWallet([
+        new WalletInterface(this.actualPrivKey, false, privKeyType)
+      ]).then(() => {
+        this.actualPrivKey = '';
+        this.spinner = false;
+        this.$router.push({
+          path: 'interface'
         });
+      });
     },
     focusInput() {
       this.$refs.privateKeyInput.focus();